+++ src/components/content/content.ts
@@ -15,7 +15,6 @@ export class ContentWorkshop {
         private readonly viewManager: ViewManager,
         private readonly httpClient: HttpClient,
         private readonly authenticator: IAuthenticator,
-        private provisioningService: ProvisionService
     ) {
         this.viewManager = viewManager;
     }
@@ -40,23 +39,5 @@ export class ContentWorkshop {
             }
         }
         this.viewManager.openViewAsWorkshop(view);
-        try {
-            this.viewManager.notifySuccess("Website reset", `The website is being resetted...`);
-            await this.provisioningService.cleanup();
-            await this.provisioningService.provision();
-            this.viewManager.closeWorkshop("content-workshop");
-            const toast = this.viewManager.addToast("Website reset", `Website has been reset successfully...`, [
-                {
-                    title: "Reset",
-                    iconClass: "paperbits-check-2",
-                    action: async (): Promise<void> => {
-                        this.viewManager.removeToast(toast);
-                    }
-                },
-            ]);
-        } 
-        catch (error) {
-            this.viewManager.notifyError("Confirm", `Unable to reset website. Please try again later.`);
-        }
     }
 }
