+++ src/components/content/resetDetails.ts
@@ -1,7 +1,8 @@
 import * as ko from "knockout";
 import template from "./resetDetails.html";
-import { Component, Param, Event, OnMounted } from "@paperbits/common/ko/decorators";
+import { Component } from "@paperbits/common/ko/decorators";
 import { ProvisionService } from "../../services/provisioningService";
+import { ViewManager } from "@paperbits/common/ui";
 
 @Component({
     selector: "reset-details-workshop",
@@ -13,6 +14,7 @@ export class ResetDetailsWorkshop {
     public readonly response: ko.Observable<string>;
     public readonly isYes: ko.Computed<boolean>;
     constructor (
+        private readonly viewManager: ViewManager,
         private provisioningService: ProvisionService,
     ) {
         this.response = ko.observable("");
@@ -20,7 +22,23 @@ export class ResetDetailsWorkshop {
     }
 
     public async reset(): Promise<void> {
-        await this.provisioningService.cleanup();
-        await this.provisioningService.provision();
+        try {
+            this.viewManager.notifySuccess("Website reset", `The website is being resetted...`);
+            await this.provisioningService.cleanup();
+            await this.provisioningService.provision();
+            this.viewManager.closeWorkshop("content-workshop");
+            const toast = this.viewManager.addToast("Website reset", `Website has been reset successfully...`, [
+                {
+                    title: "Reset",
+                    iconClass: "paperbits-check-2",
+                    action: async (): Promise<void> => {
+                        this.viewManager.removeToast(toast);
+                    }
+                },
+            ]);
+        } 
+        catch (error) {
+            this.viewManager.notifyError("Confirm", `Unable to reset website. Please try again later.`);
+        }
     }
 }
